Reorganize repo, add placeholder for init()
