+++ packages/astro/package.json
@@ -69,9 +69,9 @@
     "shiki": "^0.9.10",
     "shorthash": "^0.0.2",
     "slash": "^4.0.0",
-    "srcset-parse": "^1.1.0",
     "source-map": "^0.7.3",
     "sourcemap-codec": "^1.4.8",
+    "srcset-parse": "^1.1.0",
     "string-width": "^5.0.0",
     "strip-ansi": "^7.0.1",
     "supports-esm": "^1.0.0",
@@ -91,7 +91,7 @@
     "@types/yargs-parser": "^20.2.1",
     "chai": "^4.3.4",
     "cheerio": "^1.0.0-rc.10",
-    "mocha": "^9.1.1"
+    "mocha": "^9.1.2"
   },
   "engines": {
     "node": "^12.20.0 || ^14.13.1 || >=16.0.0",
