+++ packages/astro/src/ssr/html.ts
@@ -1,6 +1,4 @@
-import type { AstroComponentMetadata } from '../@types/astro.js';
-
-import { h } from './h.js';
+import { h } from '../runtime/h.js';
 
 async function renderToStaticMarkup(tag: string, props: Record<string, any>, children: string | undefined) {
   const html = await h(tag, props, Promise.resolve(children));
