+++ packages/astro/src/ssr/rss.ts
@@ -1,4 +1,5 @@
 import type { RSSFunction, RSSFunctionArgs, RSSResult, RouteData } from '../@types/astro';
+
 import parser from 'fast-xml-parser';
 import { canonicalURL } from './util.js';
