+++ packages/astro/src/ssr/util.ts
