+++ packages/astro/src/ssr/vite/config.ts
