+++ packages/astro/src/ssr/vite/plugin-astro.ts
@@ -1,6 +1,6 @@
 import type { TransformResult } from '@astrojs/compiler';
 import type { Plugin } from 'vite';
-import type { AstroConfig, Renderer } from '../../@types/astro.js';
+import type { AstroConfig } from '../../@types/astro.js';
 
 import esbuild from 'esbuild';
 import fs from 'fs';
