+++ packages/astro/src/ssr/vite/plugin-jsx.ts
