+++ packages/astro/src/ssr/vite/plugin-markdown.ts
@@ -1,5 +1,5 @@
 import type { Plugin } from 'vite';
-import type { AstroConfig, Renderer } from '../../@types/astro.js';
+import type { AstroConfig } from '../../@types/astro.js';
 
 import esbuild from 'esbuild';
 import fs from 'fs';
